requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
async-timeout==5.0.1
attrs==25.1.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
distlib==0.3.9
distro==1.9.0
exceptiongroup==1.2.2
fastapi==0.115.8
filelock==3.18.0
frozenlist==1.5.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
idna==3.10
jiter==0.9.0
multidict==6.1.0
openai==1.70.0
platformdirs==4.3.7
propcache==0.2.1
psycopg2-binary==2.9.9
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.10.1
PyPDF2==3.0.1
python-dotenv==1.0.1
python-multipart==0.0.20
requests==2.32.3
sniffio==1.3.1
SQLAlchemy==2.0.25
starlette==0.45.3
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
virtualenv==20.29.3
yarl==1.18.3
torch>=2.0.0
transformers>=4.30.0
numpy>=1.24.0
anthropic==0.18.1 
supabase
